requirement.txt
flask
psycopg2-binary
flask_sqlalchemy
sqlalchemy_utils
passlib
openpyxl
flask_weasyprint
python-dotenv
flask-migrate
